move tugboat errors
