+++ ui/src/tugboat/components/SelectExample.tsx
@@ -4,20 +4,7 @@ import { Form } from 'antd';
 import { Field, Select } from './form';
 import { Examples } from '../context';
 import { Example, GroupedExamples, isGroupedExamples } from '../lib';
-
-class DuplicateDisplayPropValueError extends Error {
-    constructor(propName: string, value: string) {
-        super(
-            `The ${propName} property isn't unique, "${value}" is duplicated in several examples.`
-        );
-    }
-}
-
-class InvalidDisplayPropError extends Error {
-    constructor(prop: string) {
-        super(`No property named ${prop} exists on the loaded examples.`);
-    }
-}
+import { InvalidDisplayPropError, DuplicateDisplayPropValueError } from '../error';
 
 function flattenExamples(examples: GroupedExamples | Example[]): Example[] {
     if (!isGroupedExamples(examples)) {
