+++ ui/src/tugboat/error/index.ts
@@ -1,3 +1,4 @@
+// Model errors
 export class NoModelsError extends Error {
     constructor() {
         super('No models.');
@@ -28,18 +29,21 @@ export class InvalidModelResponseError extends Error {
     }
 }
 
+// Action errors
 export class UnknownActionError extends Error {
     constructor(actionName: string) {
         super(`Unknown action: ${actionName}`);
     }
 }
 
+// State errors
 export class UnknownStateError extends Error {
     constructor(state: any) {
         super(`Unknown state: ${state.constructor.name}`);
     }
 }
 
+// Example errors
 export class UngroupedExamplesError extends Error {
     constructor() {
         super(`The examples aren't grouped but are expected to be.`);
@@ -49,5 +53,42 @@ export class UngroupedExamplesError extends Error {
 export class GroupedExamplesError extends Error {
     constructor() {
         super(`The examples are grouped but are expected not to be.`);
+    }
+}
+
+export class DuplicateDisplayPropValueError extends Error {
+    constructor(propName: string, value: string) {
+        super(
+            `The ${propName} property isn't unique, "${value}" is duplicated in several examples.`
+        );
+    }
+}
+
+export class InvalidDisplayPropError extends Error {
+    constructor(prop: string) {
+        super(`No property named ${prop} exists on the loaded examples.`);
+    }
+}
+
+// Config errors
+export class ConfigError extends Error {
+    constructor(message: string) {
+        super(`Configuration Error: ${message}`);
+    }
+}
+
+// Form errors
+export class EmptyFormError extends Error {
+    constructor() {
+        super(
+            'The form was submitted while empty. This is indicative of a programming error and should not happen'
+        );
+    }
+}
+
+// Highlight errors
+export class InvalidHighlightRangeError extends Error {
+    constructor(msg: string) {
+        super(`Invalid Highlight Range: ${msg}`);
     }
 }
